cleanup and restructured files
