+++ src/components/MainTiles/MainTilesStyles.js
