+++ src/index.js
@@ -2,6 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import 'normalize.css';
-import MainTiles from './components/MainTiles.jsx'
+import MainTiles from './components/MainTiles/MainTiles'
 
 ReactDOM.render(<MainTiles />, document.getElementById('root'));
